+++ src/Factory/AuthorizeMiddlewareFactory.php
@@ -7,7 +7,7 @@ namespace Sirix\Mezzio\Rbac\Factory;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
-use Sirix\Mezzio\Rbac\Contract\GuardInterface;
+use Sirix\Mezzio\Rbac\Contract\RequestGuardInterface;
 use Sirix\Mezzio\Rbac\Middleware\AuthorizeMiddleware;
 
 final class AuthorizeMiddlewareFactory
@@ -19,7 +19,7 @@ final class AuthorizeMiddlewareFactory
     public function __invoke(ContainerInterface $container): AuthorizeMiddleware
     {
         return new AuthorizeMiddleware(
-            $container->get(GuardInterface::class),
+            $container->get(RequestGuardInterface::class),
         );
     }
 }