+++ src/Factory/RequestGuardFactory.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sirix\Mezzio\Rbac\Factory;
+
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
+use Sirix\Mezzio\Rbac\AuthorizationEvaluator;
+use Sirix\Mezzio\Rbac\Contract\RequestActorProviderInterface;
+use Sirix\Mezzio\Rbac\Contract\RequestGuardInterface;
+use Sirix\Mezzio\Rbac\RequestGuard;
+
+final class RequestGuardFactory
+{
+    /**
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function __invoke(ContainerInterface $container): RequestGuardInterface
+    {
+        return new RequestGuard(
+            $container->get(RequestActorProviderInterface::class),
+            $container->get(AuthorizationEvaluator::class),
+        );
+    }
+}